Renamed Project to QOL-iOS-Library for more professionalism. Fixed all of the warnings.

+++ QOL-iOS-Library/Categories/NSString/NSString+Formatting.h
@@ -26,7 +26,7 @@
 - (NSString*)httpAddress;
 
 /*! returns the port. Defaults to 80 if none found*/
-- (unsigned int)port;
+- (NSInteger)port;
 
 /*! This method will retain the subpath */
 - (NSString*)httpAddressWithSubpath;
